fix: move kubectl related util to watch events
