+++ tools/ark-cli/src/lib/constants.ts
@@ -8,3 +8,9 @@ export const QUERY_ANNOTATIONS = {
   A2A_CONTEXT_ID: `${ARK_PREFIX}a2a-context-id`,
 } as const;
 
+// Event annotation constants
+export const EVENT_ANNOTATIONS = {
+  // Event data annotation for structured event data
+  EVENT_DATA: `${ARK_PREFIX}event-data`,
+} as const;
+
