+++ tools/ark-cli/src/lib/kubectl.ts
@@ -1,5 +1,7 @@
 import {execa} from 'execa';
+import chalk from 'chalk';
 import type {K8sListResource} from './types.js';
+import {EVENT_ANNOTATIONS} from './constants.js';
 
 interface K8sResource {
   metadata: {
@@ -91,4 +93,84 @@ export async function replaceResource<T extends K8sResource>(
   });
 
   return JSON.parse(result.stdout) as T;
+}
+
+export async function watchEventsLive(queryName: string): Promise<void> {
+  const seenEvents = new Set<string>();
+
+  const pollEvents = async () => {
+    try {
+      const {stdout} = await execa('kubectl', [
+        'get',
+        'events',
+        '--field-selector',
+        `involvedObject.name=${queryName}`,
+        '-o',
+        'json',
+      ]);
+
+      const eventsData = JSON.parse(stdout);
+      for (const event of eventsData.items || []) {
+        const eventId = event.metadata?.uid;
+
+        if (eventId && !seenEvents.has(eventId)) {
+          seenEvents.add(eventId);
+
+          const annotations = event.metadata?.annotations || {};
+          const eventData = annotations[EVENT_ANNOTATIONS.EVENT_DATA];
+
+          if (eventData) {
+            const now = new Date();
+            const hours = now.getHours().toString().padStart(2, '0');
+            const minutes = now.getMinutes().toString().padStart(2, '0');
+            const seconds = now.getSeconds().toString().padStart(2, '0');
+            const millis = now.getMilliseconds().toString().padStart(3, '0');
+            const timestamp = `${hours}:${minutes}:${seconds}.${millis}`;
+
+            const reason = event.reason || 'Unknown';
+            const eventType = event.type || 'Normal';
+
+            const colorCode =
+              eventType === 'Normal' ? 32 : eventType === 'Warning' ? 33 : 31;
+            console.log(
+              `${timestamp} \x1b[${colorCode}m${reason}\x1b[0m ${eventData}`
+            );
+          }
+        }
+      }
+    } catch (error) {}
+  };
+
+  const pollInterval = setInterval(pollEvents, 200);
+
+  const timeoutSeconds = 300;
+
+  const waitProcess = execa(
+    'kubectl',
+    [
+      'wait',
+      '--for=condition=Completed',
+      `query/${queryName}`,
+      `--timeout=${timeoutSeconds}s`,
+    ],
+    {
+      timeout: timeoutSeconds * 1000,
+    }
+  );
+
+  try {
+    await waitProcess;
+    await pollEvents();
+    await new Promise(resolve => setTimeout(resolve, 200));
+    await pollEvents();
+  } catch (error) {
+    console.error(
+      chalk.red(
+        'Query wait failed:',
+        error instanceof Error ? error.message : 'Unknown error'
+      )
+    );
+  } finally {
+    clearInterval(pollInterval);
+  }
 }
